Support setting any format type using the format/_format option
If set, the format/_format option will cause an appropriate format attribute to be set on the RPC.
Also, we default to not mangling the return value and only do that if the format was "dict".
Finally, fixed a typo in the cli method that would cause double invocation of RPCs when the format was not "dict".